server/requirements.txt
bcrypt==4.0.1
pyjwt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy
alembic==1.13.1
psycopg[binary]
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
cloudinary==1.40.0
python-dotenv==1.0.1
websockets==12.0
python-dotenv
